Добавление в конфиг 2 пользователей
