--- src/main/java/com/example/tacos1_3/WebConfig.java
@@ -1,37 +1,0 @@
-package com.example.tacos1_3;
-
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-//import org.springframework.data.jdbc.repository.config.EnableJdbcRepositories;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@ComponentScan(basePackages = "com.*")
-@EntityScan(basePackages = "com.model")
-//@EnableJdbcRepositories(basePackages = "com.repository")
-@EnableJpaRepositories(basePackages = "com.repository")
-public class WebConfig implements WebMvcConfigurer {
-    @Override
-    public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController("/").setViewName("home");
-    }
-
-//    @Bean
-//    public ApplicationRunner dataLoader(IngredientRepository repo) {
-//        return args -> {
-//            repo.save(new Ingredient("FLTO", "Flour Tortilla", Type.WRAP));
-//            repo.save(new Ingredient("COTO", "Corn Tortilla", Type.WRAP));
-//            repo.save(new Ingredient("GRBF", "Ground Beef", Type.PROTEIN));
-//            repo.save(new Ingredient("CARN", "Carnitas", Type.PROTEIN));
-//            repo.save(new Ingredient("TMTO", "Diced Tomatoes", Type.VEGGIES));
-//            repo.save(new Ingredient("LETC", "Lettuce", Type.VEGGIES));
-//            repo.save(new Ingredient("CHED", "Cheddar", Type.CHEESE));
-//            repo.save(new Ingredient("JACK", "Monterrey Jack", Type.CHEESE));
-//            repo.save(new Ingredient("SLSA", "Salsa", Type.SAUCE));
-//            repo.save(new Ingredient("SRCR", "Sour Cream", Type.SAUCE));
-//        };
-//    }
-}
